Pass un-parsed innerHTML into getAttribute to support raw property
In #39424 when we added an optimization to only parse a block's
innerHTML once we also changed the behavior that the innerHTML
propety represented the raw HTML loaded by the parser. Instead,
what we have since that change is the DOM of the parsed HTML.

In this patch we're adding yet-another parameter to the bag of
arguments in `getBlockAttribute()` so that the new `raw` type
can read and reproduce that original source, e.g. when reading
the string `1 < 0` the parsed value's `innerHTML` will be
`1 &lt; 0` even though the block's raw content was `1 < 0`.