Simplify and encapsulate getting provider instance
Props @aristath

The knowledge of how to get the instance and put it
into the store is now encapsulated in a private
method. Why? It's a different task from what do_item()
is doing. By moving to a private method, its intent
is documented and more clear for future understanding.